Try to work with non-async kernel clients
When testing out the WWT "widget" mode (as opposed to the JupyterLab WWT
"app"), I ran into exceptions that seem to originate in the fact that
the kernel client was apparently synchronous rather than asynchronous. I
wish I understood the overall situation better, but it looks like we can
work around the issue by using an `async def` and deciding whether we
need to `await` on the message just based on whether it's a dict or not.