Fix support for simultaneous HTTP requests
If the KDR gets a lot of requests at once, the different request
handlers can easily steal each others' messages. To fix this, we need to
make sure that when that happens, we buffer up those messages and
deliver them to their intended recipients. This gets a little bit messy
but fundamentally all we're doing is adding a buffering and dispatch
layer.