Even more multi-async fixes
It looks like we should be buffering not only sessions, but kernel
clients ... which actually makes a ton of sense. Doing this exposed
another weakness in the buffering scheme: if we ask the client for a
message and get one, someone else might still have buffered an earlier
message for us. Fix that by just always using the buffer.